Removed SuperSamp class in favor of averaging subsamples early.
SuperSamp was only really used to retain the subsamples for a given
pixel while it was passed around, before being merged averaged into a
single value. It’s easier, more readable and more efficient to merge
them immediately.

This seems to result in a nice speed bump for the first few frames, and
one fewer class to worry about.