Redefined Retracer.Radiance() as a tail-recursive function.
Raytracer.radiance() was defined as a series of mutually recursive
calls to material.radiance() and itself, and the state of the RNG was
passed along implicitly through the use of the State monad (from Cats).
It is simpler, and certainly faster and safer, to define
Raytracer.radiance() as a tail-recursive function.